conserta a busca binária e reune codigos em um único arquivo
